refactor(editor): engine final split 적용
