+++ front/e2e/editor-studio-state.spec.ts
@@ -129,7 +129,7 @@ test.describe("editor studio state", () => {
     expect(editorEditSource).toContain("const EditorPostPage: NextPage<AdminPageProps> = (props) => <EditorStudioPage {...props} />")
   })
 
-  test("editor studio는 v2 단일 경로와 단일 작성 모드 계약을 유지한다", () => {
+  test("editor studio와 BlockEditorEngine은 v2 단일 경로와 단일 작성 모드 계약을 유지한다", () => {
     const editorStudioPageSource = readFileSync(path.resolve(__dirname, "../src/routes/Admin/EditorStudioPage.tsx"), "utf8")
     const editorStudioSource = readFileSync(
       path.resolve(__dirname, "../src/routes/Admin/EditorStudioWorkspaceController.tsx"),
@@ -391,6 +391,14 @@ test.describe("editor studio state", () => {
       : ""
     const blockEditorShellSource = readFileSync(path.resolve(__dirname, "../src/components/editor/BlockEditorShell.tsx"), "utf8")
     const blockEditorEngineSource = readFileSync(path.resolve(__dirname, "../src/components/editor/BlockEditorEngine.tsx"), "utf8")
+    const blockEditorEngineControllerPath = path.resolve(
+      __dirname,
+      "../src/components/editor/useBlockEditorEngineController.ts"
+    )
+    expect(existsSync(blockEditorEngineControllerPath)).toBe(true)
+    const blockEditorEngineControllerSource = existsSync(blockEditorEngineControllerPath)
+      ? readFileSync(blockEditorEngineControllerPath, "utf8")
+      : ""
     const blockEditorEngineStylesPath = path.resolve(
       __dirname,
       "../src/components/editor/BlockEditorEngine.styles.tsx"
@@ -1081,11 +1089,13 @@ test.describe("editor studio state", () => {
     expect(blockEditorShellSource).toContain('import BlockEditorEngine from "./BlockEditorEngine"')
     expect(blockEditorShellSource).toContain('import type { BlockEditorEngineProps } from "./blockEditorEngineTypes"')
     expect(blockEditorShellSource).toContain("const BlockEditorShell = (props: BlockEditorEngineProps) => <BlockEditorEngine {...props} />")
-    expect(blockEditorEngineSource.split("\n").length).toBeLessThanOrEqual(5500)
+    expect(blockEditorEngineSource.split("\n").length).toBeLessThanOrEqual(1000)
+    expect(blockEditorEngineSource).toContain('from "./useBlockEditorEngineController"')
+    expect(blockEditorEngineControllerSource).toContain("export const useBlockEditorEngineController =")
     expect(blockEditorEngineSource).toContain('} from "./BlockEditorEngine.styles"')
     expect(blockEditorEngineSource).toContain('} from "./BlockEditorEngine.layers"')
     expect(blockEditorEngineSource).toContain('} from "./BlockEditorEngine.tableOverlayLayer"')
-    expect(blockEditorEngineSource).toContain('} from "./useBlockEditorTableOverlayController"')
+    expect(blockEditorEngineControllerSource).toContain('} from "./useBlockEditorTableOverlayController"')
     expect(blockEditorTableOverlayLayerSource).toContain("export const BlockEditorTableOverlayLayer =")
     expect(blockEditorTableOverlayControllerSource).toContain("export const useBlockEditorTableOverlayController =")
     expect(editorExtensionsSource).toContain('from "./codeBlockNodeView"')
@@ -1186,13 +1196,13 @@ test.describe("editor studio state", () => {
     expect(blockEditorEngineStylesSource).toContain(".aq-block-editor__content blockquote {")
     expect(blockEditorEngineStylesSource).toContain("border-left: 4px solid")
     expect(blockEditorEngineStylesSource).toContain("border-radius: 0;")
-    expect(blockEditorEngineSource).toContain('from "./blockSelectionModel"')
-    expect(blockEditorEngineSource).toContain('from "./nestedListItemModel"')
-    expect(blockEditorEngineSource).toContain('from "./useBlockEditorMarkdownCommit"')
-    expect(blockEditorEngineSource).toContain('from "./slashMenuModel"')
-    expect(blockEditorEngineSource).toContain('from "./blockToolbarModel"')
-    expect(blockEditorEngineSource).toContain('from "./inlineToolbarModel"')
-    expect(blockEditorEngineSource).toContain('from "./useFloatingBubbleState"')
+    expect(blockEditorEngineControllerSource).toContain('from "./blockSelectionModel"')
+    expect(blockEditorEngineControllerSource).toContain('from "./nestedListItemModel"')
+    expect(blockEditorEngineControllerSource).toContain('from "./useBlockEditorMarkdownCommit"')
+    expect(blockEditorEngineControllerSource).toContain('from "./slashMenuModel"')
+    expect(blockEditorEngineControllerSource).toContain('from "./blockToolbarModel"')
+    expect(blockEditorEngineControllerSource).toContain('from "./inlineToolbarModel"')
+    expect(blockEditorEngineControllerSource).toContain('from "./useFloatingBubbleState"')
     expect(blockEditorEngineSource).not.toContain("type FloatingBubbleState =")
     expect(blockEditorEngineSource).not.toContain("const bubbleHideTimerRef")
     expect(blockEditorEngineSource).not.toContain("const bubbleToolbarHoveredRef = useRef")
@@ -1231,19 +1241,19 @@ test.describe("editor studio state", () => {
     expect(inlineToolbarModelSource).toContain("export const runInlineCode")
     expect(inlineToolbarModelSource).toContain("export const runInlineColor")
     expect(inlineToolbarModelSource).toContain("export const runInlineTextStyle")
-    expect(blockEditorEngineSource).toContain('runInlineMarkCommand(editor, "bold")')
-    expect(blockEditorEngineSource).toContain('runInlineMarkCommand(editor, "italic")')
-    expect(blockEditorEngineSource).toContain('runInlineMarkCommand(editor, "strike")')
-    expect(blockEditorEngineSource).toContain('runBlockToolbarCommand(currentEditor, "heading-2")')
-    expect(blockEditorEngineSource).toContain('runBlockToolbarCommand(currentEditor, "heading-3")')
-    expect(blockEditorEngineSource).toContain('runBlockToolbarCommand(currentEditor, "ordered-list")')
-    expect(blockEditorEngineSource).toContain('runBlockToolbarCommand(currentEditor, "bullet-list")')
-    expect(blockEditorEngineSource).toContain('runBlockToolbarCommand(currentEditor, "quote")')
-    expect(blockEditorEngineSource).toContain('runBlockToolbarCommand(editor, "heading-1")')
-    expect(blockEditorEngineSource).toContain('runBlockToolbarCommand(editor, "bullet-list")')
-    expect(blockEditorEngineSource).toContain("isToolbarBlockInsertActive(editor, item.id)")
-    expect(blockEditorEngineSource).toContain("runInlineCode(editor)")
-    expect(blockEditorEngineSource).toContain("runInlineColor(editor, color)")
+    expect(blockEditorEngineControllerSource).toContain('runInlineMarkCommand(editor, "bold")')
+    expect(blockEditorEngineControllerSource).toContain('runInlineMarkCommand(editor, "italic")')
+    expect(blockEditorEngineControllerSource).toContain('runInlineMarkCommand(editor, "strike")')
+    expect(blockEditorEngineControllerSource).toContain('runBlockToolbarCommand(currentEditor, "heading-2")')
+    expect(blockEditorEngineControllerSource).toContain('runBlockToolbarCommand(currentEditor, "heading-3")')
+    expect(blockEditorEngineControllerSource).toContain('runBlockToolbarCommand(currentEditor, "ordered-list")')
+    expect(blockEditorEngineControllerSource).toContain('runBlockToolbarCommand(currentEditor, "bullet-list")')
+    expect(blockEditorEngineControllerSource).toContain('runBlockToolbarCommand(currentEditor, "quote")')
+    expect(blockEditorEngineControllerSource).toContain('runBlockToolbarCommand(editor, "heading-1")')
+    expect(blockEditorEngineControllerSource).toContain('runBlockToolbarCommand(editor, "bullet-list")')
+    expect(blockEditorEngineControllerSource).toContain("isToolbarBlockInsertActive(editor, item.id)")
+    expect(blockEditorEngineControllerSource).toContain("runInlineCode(editor)")
+    expect(blockEditorEngineControllerSource).toContain("runInlineColor(editor, color)")
     expect(blockEditorEngineSource).not.toContain("toggleBold().run()")
     expect(blockEditorEngineSource).not.toContain("toggleItalic().run()")
     expect(blockEditorEngineSource).not.toContain("toggleStrike().run()")
@@ -1285,8 +1295,8 @@ test.describe("editor studio state", () => {
     expect(blockEditorEngineSource).not.toMatch(/const\s+getEditableTextPositionForTopLevelBlock\s*=/)
     expect(blockEditorEngineSource).not.toMatch(/const\s+selectTopLevelBlockNode\s*=/)
     expect(blockEditorEngineSource).not.toMatch(/const\s+isTabBlockSelectionEligible\s*=/)
-    expect(blockEditorEngineSource).toContain('from "./blockHandleLayoutModel"')
-    expect(blockEditorEngineSource).toContain('from "./blockDragModel"')
+    expect(blockEditorEngineControllerSource).toContain('from "./blockHandleLayoutModel"')
+    expect(blockEditorEngineControllerSource).toContain('from "./blockDragModel"')
     expect(blockEditorTableOverlayControllerSource).toContain('from "./tableAxisDragModel"')
     expect(blockEditorEngineSource).not.toMatch(/const\s+resolveBlockHandleAnchorTop\s*=/)
     expect(blockEditorEngineSource).not.toMatch(/const\s+resolveThinBlockHandleAnchorTop\s*=/)
@@ -1709,8 +1719,8 @@ test.describe("editor studio state", () => {
     expect(tableWidthModelSource).toContain("export const promotePastedWideTables = (")
     expect(tableWidthModelSource).toContain("export const computeNextTableColumnWidthsForResize = (")
     expect(tableWidthModelSource).toContain("export const didTableColumnResizeHitOverflowPolicy = (")
-    expect(blockEditorEngineSource).toContain('} from "./tableWidthModel"')
-    expect(blockEditorEngineSource).toContain('} from "./tableWidthRuntime"')
+    expect(blockEditorEngineControllerSource).toContain('} from "./tableWidthModel"')
+    expect(blockEditorEngineControllerSource).toContain('} from "./tableWidthRuntime"')
     expect(tableWidthRuntimeSource).toContain("export type TableWidthSnapshot = {")
     expect(tableWidthRuntimeSource).toContain("export const getCurrentEditorReadableWidthPx = (")
     expect(tableWidthRuntimeSource).toContain("export const rebalanceStructurallyChangedNormalTableWidths = (")
@@ -1738,7 +1748,7 @@ test.describe("editor studio state", () => {
     expect(tableStructureModelSource).toContain("export const hasTableContextInResolvedPos = (")
     expect(tableStructureModelSource).toContain("export const isTableSelectionActive = (")
     expect(tableStructureModelSource).toContain("export const getActiveTableStructureState = (")
-    expect(blockEditorEngineSource).toContain('} from "./tableStructureModel"')
+    expect(blockEditorEngineControllerSource).toContain('} from "./tableStructureModel"')
     expect(blockEditorEngineSource).not.toContain("const collectSimpleTableColumnCells = (")
     expect(blockEditorEngineSource).not.toContain("const buildReorderedSimpleTableNode = (")
     expect(blockEditorEngineSource).not.toContain("const canShrinkTableAxisAtEnd = (")
@@ -1752,7 +1762,7 @@ test.describe("editor studio state", () => {
     expect(tablePasteModelSource).toContain("export const normalizeTableContextPasteText = (")
     expect(tablePasteModelSource).toContain("const normalizeTableContextPasteLine = (")
     expect(tablePasteModelSource).toContain("const isMarkdownTableRow = (")
-    expect(blockEditorEngineSource).toContain('from "./tablePasteModel"')
+    expect(blockEditorEngineControllerSource).toContain('from "./tablePasteModel"')
     expect(blockEditorEngineSource).not.toContain("const normalizeTableContextPasteText = (")
     expect(blockEditorEngineSource).not.toContain("const normalizeTableContextPasteLine = (")
     expect(blockEditorEngineSource).not.toContain("const isMarkdownTableRow = (")
